content-security-policy/to-port-from-blink/1.1/referrer-empty-https-https-expected.txt

CONSOLE ERROR: The 'referrer' Content Security Policy directive has the invalid value "". Valid values are "no-referrer", "no-referrer-when-downgrade", "origin", and "unsafe-url". Note that "origin-when-cross-origin" is not yet supported.
Test that '' == never works correctly from HTTPS to HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is ""
PASS successfullyParsed is true

TEST COMPLETE
